requirements.txt

Flask
PyPDF2
python-docx
textstat
